Add createTask function via API

+++ pom/services/task.ts
@@ -1,6 +1,7 @@
 import { AUTH_TOKEN } from "../data/Constants";
 import { TASKS_URL } from "../data/urls";
 import { type Page } from "@playwright/test";
+import { type Task } from "../data/Interfaces";
 
 export class TaskService {
 	readonly page: Page;
@@ -27,11 +28,30 @@ export class TaskService {
 	}
 
 	async deleteTask(taskId: string) {
+		if (!taskId) {
+			throw new Error("Task Id is required to delete a task.");
+		}
 		await this.page.request.delete(`${TASKS_URL}/${taskId}`, {
 			headers: {
 				Authorization: `Bearer ${AUTH_TOKEN.AUTH_TOKEN}`,
 				"Content-Type": "application/json",
 			},
 		});
+	}
+
+	async createTaskViaAPI(options: Task) {
+		try {
+			const response = await this.page.request.post(`${TASKS_URL}`, {
+				headers: {
+					Authorization: `Bearer ${AUTH_TOKEN.AUTH_TOKEN}`,
+					"Content-Type": "application/json",
+				},
+				data: options,
+			});
+			const responseData = await response.json();
+			return responseData;
+		} catch (error) {
+			throw new Error(`Error creating task via API:`, { cause: error });
+		}
 	}
 }
